Apply the new test style
