+++ examples/serve_one.rs
@@ -5,15 +5,14 @@
 //! $
 //! ```
 
-extern crate msql_srv;
-extern crate mysql;
-
-use msql_srv::*;
 use std::io;
 use std::net;
 use std::thread;
 
+use msql_srv::*;
+
 struct Backend;
+
 impl<W: io::Write> MysqlShim<W> for Backend {
     type Error = io::Error;
