+++ src/packet.rs
@@ -1,8 +1,9 @@
-use byteorder::{ByteOrder, LittleEndian};
 use std::io;
 use std::io::prelude::*;
 
-const U24_MAX: usize = 16_777_215;
+use byteorder::{ByteOrder, LittleEndian};
+
+pub const U24_MAX: usize = 16_777_215;
 
 pub struct PacketWriter<W> {
     to_write: Vec<u8>,
@@ -184,6 +185,7 @@ impl<'a> AsRef<[u8]> for Packet<'a> {
 }
 
 use std::ops::Deref;
+
 impl<'a> Deref for Packet<'a> {
     type Target = [u8];
     fn deref(&self) -> &Self::Target {
@@ -191,7 +193,7 @@ impl<'a> Deref for Packet<'a> {
     }
 }
 
-fn packet(i: &[u8]) -> nom::IResult<&[u8], (u8, Packet<'_>)> {
+pub(crate) fn packet(i: &[u8]) -> nom::IResult<&[u8], (u8, Packet<'_>)> {
     nom::combinator::map(
         nom::sequence::pair(
             nom::multi::fold_many0(
@@ -222,58 +224,4 @@ fn packet(i: &[u8]) -> nom::IResult<&[u8], (u8, Packet<'_>)> {
             (seq, pkt)
         },
     )(i)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_one_ping() {
-        assert_eq!(
-            onepacket(&[0x01, 0, 0, 0, 0x10]).unwrap().1,
-            (0, &[0x10][..])
-        );
-    }
-
-    #[test]
-    fn test_ping() {
-        let p = packet(&[0x01, 0, 0, 0, 0x10]).unwrap().1;
-        assert_eq!(p.0, 0);
-        assert_eq!(&*p.1, &[0x10][..]);
-    }
-
-    #[test]
-    fn test_long_exact() {
-        let mut data = vec![0xff, 0xff, 0xff, 0];
-        data.extend(&[0; U24_MAX][..]);
-        data.push(0x00);
-        data.push(0x00);
-        data.push(0x00);
-        data.push(1);
-
-        let (rest, p) = packet(&data[..]).unwrap();
-        assert!(rest.is_empty());
-        assert_eq!(p.0, 1);
-        assert_eq!(p.1.len(), U24_MAX);
-        assert_eq!(&*p.1, &[0; U24_MAX][..]);
-    }
-
-    #[test]
-    fn test_long_more() {
-        let mut data = vec![0xff, 0xff, 0xff, 0];
-        data.extend(&[0; U24_MAX][..]);
-        data.push(0x01);
-        data.push(0x00);
-        data.push(0x00);
-        data.push(1);
-        data.push(0x10);
-
-        let (rest, p) = packet(&data[..]).unwrap();
-        assert!(rest.is_empty());
-        assert_eq!(p.0, 1);
-        assert_eq!(p.1.len(), U24_MAX + 1);
-        assert_eq!(&p.1[..U24_MAX], &[0; U24_MAX][..]);
-        assert_eq!(&p.1[U24_MAX..], &[0x10]);
-    }
 }
