+++ src/tests/mod.rs
@@ -1,0 +1,3 @@
+mod commands;
+mod packet;
+mod value;
