+++ src/tests/value/mod.rs
@@ -1,0 +1,2 @@
+mod decode;
+mod encode;
