+++ src/value/mod.rs
@@ -1,8 +1,5 @@
 mod decode;
 mod encode;
 
-#[cfg(test)]
-pub mod utils;
-
 pub use self::decode::{Value, ValueInner};
 pub use self::encode::ToMysqlValue;
